feat: update multicast-dns dependency
Update the multicast-dns dependency to ^7.0.1. The main difference is the use
of dns-packet to encode and decode UDP messages. The dns-packet module parses
the txt field as an array. Also using the @lgs-innovations version of
multicast-dns.

BREAKING CHANGE: the txt field of the service object is parsed as an array by
  dns-packet.